+++ src/main/java/org/rcsb/cif/model/IntColumn.java
@@ -32,6 +32,8 @@ public interface IntColumn extends Column<int[]> {
         if (text.isEmpty() || ".".equals(text) || "?".equals(text)) {
             return 0;
         }
-        return Integer.parseInt(text);
+        // some floats may omit decimal places and can be parsed as int: 88. - ignore the dot (in agreement with Mol*)
+        int index = text.indexOf('.');
+        return Integer.parseInt(index == -1 ? text : text.substring(0, index));
     }
 }